pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caregiver-r1"
version = "0.1.0"
description = "EC-MTRL: GRPO with EC-MTRL extensions for dementia-care multi-turn dialogue (NeurIPS 2026 submission)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Caregiver-R1 Team" }]

dependencies = [
    # Core ML
    "torch>=2.4",
    "transformers>=4.51",          # Qwen3 thinking-mode support landed in 4.51
    "tokenizers>=0.20",
    "accelerate>=0.34",
    "datasets>=2.20",
    "numpy>=1.26",

    # Inference / rollout
    "vllm>=0.8.5",                 # >=0.8.5 has Qwen3 native backend (<0.8.5 falls back to slow Transformers backend); modern installs land on 0.19+
    "openai>=1.40",                # used as a thin client for vLLM-served judge / DemMA endpoints

    # RL framework (verl is installed separately from source - see README;
    # we depend on its runtime requirements here)
    "ray[default]>=2.40",
    "hydra-core>=1.3",
    "omegaconf>=2.3",

    # Schema / config / IO
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "jsonlines>=4.0",

    # Logging
    "wandb>=0.18",
    "rich>=13.7",

    # Statistics / evaluation
    "scipy>=1.11",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.urls]
"Method" = "see ../final_rps_method_section_locked_zh_9.md"
"Plan"   = "ENGINEERING_PLAN.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
